add a prompt:none scope with special behaviour
Recognise a special scope, defaulting to "prompt:none" but configurable
via the CONSENT_PROMPT_NONE_SCOPE setting, which will never redirect the
user to a login page but will accept or reject the consent request based
entirely on the existing login state.

This can be used to implement UI-less login flows where we can detect if
the user can be automatically logged in and, if so, log them in.

Closes #4