update docker-compose configuration
Port the ./compose.sh-style docker-compose configuration to this
application and update the instructions on getting started. This commit
does not alter the CI configuration.

This commit is a forerunner for making a Hydra version 1 compatible
flavour of the application.

We experiment with a slightly nicer way to auto-create OAuth2 clients by
having the configuration for each client in a JSON file within
compose/oauth2-clients/. See the README in that directory for more
information.